requirements.txt
Flask==3.0.0
Flask-SQLAlchemy
psycopg2-binary
Werkzeug
gunicorn
Flask-SocketIO==5.5.1
eventlet==0.33.3
requests
websocket-client
redis
rq
Pillow